Added individual pages and folder creation
