+++ .env.example
@@ -2,4 +2,4 @@ AUTH_SECRET="<openssl rand -hex 32>"
 AUTH_KEYCLOAK_ID="<KeycloakClientID>"
 AUTH_KEYCLOAK_SECRET="<KeycloakClientSecret>"
 AUTH_KEYCLOAK_ISSUER="http://<KeycloakURL>/realms/<realm name>"
-PUBLIC_ROOT_FOLDER_LOCATION="<Full path to the KeyChallenge4 folder (no slash at end)>"
+PUBLIC_ROOT_FOLDER_LOCATION="<Full path to the root folder (no slash at end)>"
