+++ .gitignore
@@ -9,4 +9,5 @@ vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
 /exampleData
 /sampleData
+/diagnostics
 KeyChallenge4
