+++ src/routes/api/save-json/+server.ts
@@ -1,6 +1,6 @@
 import { json } from '@sveltejs/kit';
 import type { RequestHandler } from './$types';
-import { writeFile, access, constants } from 'fs/promises';
+import { writeFile, access, constants, mkdir } from 'fs/promises';
 import { join, resolve, relative, normalize, basename, dirname, extname } from 'path';
 import { PUBLIC_ROOT_FOLDER_LOCATION } from '$env/static/public';
 
@@ -19,11 +19,6 @@ export const POST: RequestHandler = async ({ request }) => {
     const targetPath = normalize(jsonData.targetPath);
     const absoluteTargetPath = resolve(rootFolder, targetPath);
 
-    // Check if the target path is within the root folder
-    if (!absoluteTargetPath.startsWith(resolve(rootFolder))) {
-      return json({ success: false, message: 'Invalid target path' }, { status: 403 });
-    }
-
     // Determine the filename and directory
     let fileName, directory;
     if (extname(absoluteTargetPath).toLowerCase() === '.json') {
@@ -34,10 +29,20 @@ export const POST: RequestHandler = async ({ request }) => {
       directory = absoluteTargetPath;
     }
 
+    // Create the directory if it doesn't exist
+    try {
+      await access(directory, constants.F_OK);
+    } catch (error) {
+      // Directory doesn't exist, create it
+      await mkdir(directory, { recursive: true });
+      console.log(`Created directory: ${directory}`);
+    }
+
+    // Check if the directory is writable
     try {
       await access(directory, constants.W_OK);
     } catch (error) {
-      return json({ success: false, message: 'Target directory does not exist or is not writable' }, { status: 403 });
+      return json({ success: false, message: 'Target directory is not writable' }, { status: 403 });
     }
 
     const filePath = join(directory, fileName);
